Restructuring the PanelManager
