+++ packages/terra-clinical-application-site/src/demo/panel-manager/reducers.js
@@ -1,13 +1,13 @@
 import { open, push, pop, maximize, minimize, defaultState } from 'terra-clinical-modal-manager/lib/shared/disclosureReducerUtils';
 
 import {
-  OPEN_PANEL,
-  CLOSE_PANEL,
-  PUSH_PANEL,
-  POP_PANEL,
-  MAXIMIZE_PANEL,
-  MINIMIZE_PANEL,
-} from '../actions/panelController';
+  OPEN,
+  CLOSE,
+  PUSH,
+  POP,
+  MAXIMIZE,
+  MINIMIZE,
+} from './actionTypes';
 
 const supportedSizes = {
   tiny: 'tiny',
@@ -28,20 +28,20 @@ const defaultPanelState = Object.assign({}, defaultState, {
 });
 const panelManager = (state = defaultPanelState, action) => {
   switch (action.type) {
-    case OPEN_PANEL:
+    case OPEN:
       return Object.assign({}, open(state, action), {
         size: action.data.size || supportedSizes.small,
         behavior: action.data.behavior || supportedBehaviors.squish,
       });
-    case CLOSE_PANEL:
+    case CLOSE:
       return defaultPanelState;
-    case PUSH_PANEL:
+    case PUSH:
       return push(state, action);
-    case POP_PANEL:
+    case POP:
       return pop(state, action);
-    case MAXIMIZE_PANEL:
+    case MAXIMIZE:
       return maximize(state, action);
-    case MINIMIZE_PANEL:
+    case MINIMIZE:
       return minimize(state, action);
     default:
       return state;
